+++ tests/testthat/test_1VCFFilesToZipFile.R
@@ -159,9 +159,16 @@ test_that("VCFsToZipFile function for Mutect VCFs with no base.filename", {
                        zipfile = paste0(tempdir(), "/test.zip"), 
                        ref.genome = "hg19", variant.caller = "mutect",
                        region = "genome")
+  files <- list.files(path = dir, full.names = TRUE)
+  out1 <- VCFsToZipFile(files = files, 
+                        zipfile = paste0(tempdir(), "/test1.zip"), 
+                        ref.genome = "hg19", variant.caller = "mutect",
+                        region = "genome")
+  expect_equal(out, out1)
+  
   expect_type(out, "list")
   name <- grep("\\.zip$", list.files(tempdir()), value = TRUE)
-  expect_equal(name, "test.zip")
+  expect_equal(name, c("test.zip", "test1.zip"))
   zip::unzip(zipfile = paste0(tempdir(), "/test.zip"), exdir = tempdir())
   name1 <- grep("\\.csv$", list.files(tempdir()), value = TRUE)
   expect_equal(length(name1), 7)
@@ -169,6 +176,7 @@ test_that("VCFsToZipFile function for Mutect VCFs with no base.filename", {
   expect_equal(length(name2), 8)
   
   unlink(file.path(tempdir(), "test.zip"))
+  unlink(file.path(tempdir(), "test1.zip"))
   unlink(file.path(tempdir(), name1))
   unlink(file.path(tempdir(), name2))
   graphics.off()
@@ -184,9 +192,18 @@ test_that("VCFsToZipFile function for Mutect VCFs with base.filename", {
                        ref.genome = "hg19",
                        region = "genome", variant.caller = "mutect",
                        base.filename = "test")
+  files <- list.files(path = dir, full.names = TRUE)
+  out1 <- VCFsToZipFile(files = files, 
+                        zipfile = paste0(tempdir(), "/test1.zip"), 
+                        ref.genome = "hg19",
+                        region = "genome", variant.caller = "mutect",
+                        base.filename = "test")
+  expect_equal(out, out1)
+  
+  
   expect_type(out, "list")
   name <- grep("\\.zip$", list.files(tempdir()), value = TRUE)
-  expect_equal(name, "test.zip")
+  expect_equal(name, c("test.zip", "test1.zip"))
   zip::unzip(zipfile = paste0(tempdir(), "/test.zip"), exdir = tempdir())
   name1 <- grep("\\.csv$", list.files(tempdir()), value = TRUE)
   expect_equal(length(name1), 7)
@@ -208,9 +225,17 @@ test_that("VCFsToZipFile function for Strelka SBS VCFs with no base.filename", {
                        zipfile = paste0(tempdir(), "/test.zip"), 
                        ref.genome = "hg19", variant.caller = "strelka",
                        region = "genome")
+  
+  files <- list.files(path = dir, full.names = TRUE)
+  out1 <- VCFsToZipFile(files = files,
+                        zipfile = paste0(tempdir(), "/test1.zip"), 
+                        ref.genome = "hg19", variant.caller = "strelka",
+                        region = "genome")
+  expect_equal(out, out1)
+  
   expect_type(out, "list")
   name <- grep("\\.zip$", list.files(tempdir()), value = TRUE)
-  expect_equal(name, "test.zip")
+  expect_equal(name, c("test.zip", "test1.zip"))
   zip::unzip(zipfile = paste0(tempdir(), "/test.zip"), exdir = tempdir())
   name1 <- grep("\\.csv$", list.files(tempdir()), value = TRUE)
   expect_equal(length(name1), 7)
@@ -233,9 +258,18 @@ test_that("VCFsToZipFile function for Strelka SBS VCFs with base.filename", {
                        ref.genome = "hg19",
                        region = "genome", variant.caller = "strelka",
                        base.filename = "test")
+  
+  files <- list.files(path = dir, full.names = TRUE)
+  out1 <- VCFsToZipFile(files = files,
+                        zipfile = paste0(tempdir(), "/test1.zip"), 
+                        ref.genome = "hg19",
+                        region = "genome", variant.caller = "strelka",
+                        base.filename = "test")
+  expect_equal(out, out1)
+  
   expect_type(out, "list")
   name <- grep("\\.zip$", list.files(tempdir()), value = TRUE)
-  expect_equal(name, "test.zip")
+  expect_equal(name, c("test.zip", "test1.zip"))
   zip::unzip(zipfile = paste0(tempdir(), "/test.zip"), exdir = tempdir())
   name1 <- grep("\\.csv$", list.files(tempdir()), value = TRUE)
   expect_equal(length(name1), 7)
@@ -257,9 +291,17 @@ test_that("VCFsToZipFile function for Strelka ID VCFs with no base.filename", {
                        zipfile = paste0(tempdir(), "/test.zip"),
                        ref.genome = "hg19", variant.caller = "strelka",
                        region = "genome")
+  
+  files <- list.files(path = dir, full.names = TRUE)
+  out1 <- VCFsToZipFile(files = files,
+                        zipfile = paste0(tempdir(), "/test.zip"),
+                        ref.genome = "hg19", variant.caller = "strelka",
+                        region = "genome")
+  expect_equal(out, out1)
+  
   expect_type(out, "list")
   name <- grep("\\.zip$", list.files(tempdir()), value = TRUE)
-  expect_equal(name, "test.zip")
+  expect_equal(name, c("test.zip", "test1.zip"))
   
   zip::unzip(zipfile = paste0(tempdir(), "/test.zip"), exdir = tempdir())
   name1 <- grep("\\.csv$", list.files(tempdir()), value = TRUE)
@@ -283,9 +325,18 @@ test_that("VCFsToZipFile function for Strelka ID VCFs with base.filename", {
                        ref.genome = "hg19", variant.caller = "strelka",
                        region = "genome",
                        base.filename = "test")
+  
+  files <- list.files(path = dir, full.names = TRUE)
+  out1 <- VCFsToZipFile(files = files,
+                        zipfile = paste0(tempdir(), "/test1.zip"),
+                        ref.genome = "hg19", variant.caller = "strelka",
+                        region = "genome",
+                        base.filename = "test")
+  expect_equal(out, out1)
+  
   expect_type(out, "list")
   name <- grep("\\.zip$", list.files(tempdir()), value = TRUE)
-  expect_equal(name, "test.zip")
+  expect_equal(name, c("test.zip", "test1.zip"))
   
   zip::unzip(zipfile = paste0(tempdir(), "/test.zip"), exdir = tempdir())
   name1 <- grep("\\.csv$", list.files(tempdir()), value = TRUE)